Move createAnnotations to ElementHandlerUtil
