+++ kotlinpoet-metadata-specs/src/main/kotlin/com/squareup/kotlinpoet/metadata/specs/internal/ElementHandlerUtil.kt
@@ -2,10 +2,13 @@ package com.squareup.kotlinpoet.metadata.specs.internal
 
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.AnnotationSpec.UseSiteTarget
+import com.squareup.kotlinpoet.AnnotationSpec.UseSiteTarget.FIELD
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.joinToCode
+import java.util.Collections
+import java.util.TreeSet
 
 object ElementHandlerUtil {
   internal val JVM_FIELD = JvmField::class.asClassName()
@@ -19,6 +22,32 @@ object ElementHandlerUtil {
       JVM_TRANSIENT,
       JVM_VOLATILE
   )
+
+  /**
+   * @return a new collection of [AnnotationSpecs][AnnotationSpec] with sorting and de-duping
+   *         input annotations from [body].
+   */
+  fun createAnnotations(
+      siteTarget: UseSiteTarget? = null,
+      body: MutableCollection<AnnotationSpec>.() -> Unit
+  ): Collection<AnnotationSpec> {
+    val result = TreeSet<AnnotationSpec>(compareBy { it.toString() })
+        .apply(body)
+    val withUseSiteTarget = if (siteTarget != null) {
+      result.map {
+        if (!(siteTarget == FIELD && it.className in IMPLICIT_FIELD_ANNOTATIONS)) {
+          // Some annotations are implicitly only for FIELD, so don't emit those site targets
+          it.toBuilder().useSiteTarget(siteTarget).build()
+        } else {
+          it
+        }
+      }
+    } else {
+      result
+    }
+
+    return Collections.unmodifiableCollection(withUseSiteTarget)
+  }
 
   /**
    * @return a [@Throws][Throws] [AnnotationSpec] representation of a given collection of
